+++ src/pages/create/addBillStep/components/NumPadBottomSheet/index.tsx
@@ -5,7 +5,7 @@ import {
   DrawerRoot,
   DrawerTrigger,
 } from '@/common/components/Drawer/drawer';
-import NumPad from '../NumPad.tsx';
+import NumPad from '../NumPad';
 import * as S from './index.styles';
 
 interface NumPadBottomSheetProps {
